shift-detection/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
argparse==1.4.0
attrs==25.4.0
audioread==3.1.0
automat==25.4.16
buildtools==1.0.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
docopt==0.6.2
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
furl==2.1.4
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.2
hyperlink==21.0.0
idna==3.11
imageio==2.37.2
immutabledict==4.2.2
incremental==24.7.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.9.2
llvmlite==0.45.1
markupsafe==3.0.3
matplotlib==3.9.2
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
numba==0.62.1
numpy==2.3.4
opencv-python==4.11.0.86
orderedmultidict==1.0.1
ortools==9.11.4210
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==5.26.1
pyarrow==22.0.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
redo==3.0.0
regex==2024.9.11
requests==2.32.5
resampy==0.4.3
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
sqlalchemy==2.0.44
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
twisted==25.5.0
typer-slim==0.20.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zope-interface==8.0.1
